requirements.txt
streamlit
pandas
crawl4ai
google-generativeai
groq
python-dotenv
aiohttp
nest-asyncio